Renaming centralized -> centralized_telescope
